admin/transform: allow viewing unknown transforms with committed offsets
This will be used by tests to ensure we're doing cleanup correctly.
