+++ src/v/redpanda/admin/util.cc
@@ -23,4 +23,17 @@ void apply_validator(
           "JSON request body does not conform to schema: {}", err.what()));
     }
 }
+
+bool get_boolean_query_param(
+  const ss::http::request& req, std::string_view name) {
+    auto key = ss::sstring(name);
+    if (!req.query_parameters.contains(key)) {
+        return false;
+    }
+
+    const ss::sstring& str_param = req.query_parameters.at(key);
+    return ss::internal::case_insensitive_cmp()(str_param, "true")
+           || str_param == "1";
+}
+
 } // namespace admin
